Sync fixes

+++ src/electron/cloud/syncManager.ts
@@ -5,7 +5,7 @@ import type { Folders, Projects } from "../../types/Projects"
 import type { Show } from "../../types/Show"
 import { isValidJSON, startBackup } from "../data/backup"
 import { _store, setStore } from "../data/store"
-import { compressToZip, decompressZipStream } from "../data/zip"
+import { compressToZip, decompressZipStream, getZipModifiedDates } from "../data/zip"
 import { sendMain } from "../IPC/main"
 import { createFolder, deleteFile, deleteFolder, doesPathExistAsync, getDataFolderPath, getFileStatsAsync, getTimePointString, moveFileAsync, readFileAsync, readFolderAsync, writeFileAsync } from "../utils/files"
 import { clone, getMachineId } from "../utils/helpers"
@@ -57,7 +57,7 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
     console.log("Syncing to cloud")
 
     // clear any uncleared previous data
-    deleteFolder(EXTRACT_LOCATION)
+    if (await doesPathExistAsync(EXTRACT_LOCATION)) deleteFolder(EXTRACT_LOCATION)
 
     const cloudDataPath = await provider.getData(data.churchId, data.teamId, EXTRACT_LOCATION)
     if (!cloudDataPath) {
@@ -70,6 +70,7 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
     const extractedFiles = await decompressZipStream(cloudDataPath, true, {
         getOutputPath: (fileName: string) => path.join(EXTRACT_LOCATION, fileName)
     })
+    const modifiedDates = await getZipModifiedDates(cloudDataPath)
     console.log("Files:", extractedFiles.length)
 
     const showsFolder = getDataFolderPath("shows")
@@ -79,6 +80,9 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
     if (typeof changesFile?.content === "string") {
         const changesContent = await readFileAsync(changesFile.content)
         if (isValidJSON(changesContent)) CHANGES = JSON.parse(changesContent)
+        if (CHANGES.version !== version) CHANGES = clone(DEFAULT_CHANGES)
+        cloudChanges = clone(CHANGES)
+        isNewDevice = !CHANGES.devices.find((id) => id === deviceId)
     }
     // console.log("Devices:", CHANGES.devices)
 
@@ -108,7 +112,7 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
                 }
 
                 // exists both locally and in cloud
-                if (await cloudIsNewer(cloudPath, bibleDestPath)) await download()
+                if (await cloudIsNewer(bibleDestPath, modifiedDates[file.name])) await download()
 
                 async function download() {
                     await moveFileAsync(cloudPath, bibleDestPath)
@@ -161,16 +165,12 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
                 // check any local instance not in cloud (exists only locally)
                 const showNames = await readFolderAsync(showsFolder)
                 const localShows = showNames.filter((key) => !cloudShowNames.includes(key))
-                localShows.forEach((fileName) => {
-                    if (isDeleted(id, fileName)) {
+                await Promise.all(
+                    localShows.map(async (fileName) => {
                         const localShowPath = path.join(showsFolder, fileName)
-                        deleteFile(localShowPath)
-                        return
-                    }
-
-                    // new file
-                    markAsCreated(id, fileName)
-                })
+                        await createOrDeleteLocalFile(id, fileName, localShowPath)
+                    })
+                )
                 return
             }
 
@@ -185,7 +185,7 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
             // replace full files (settings)
             if (!MERGE_INDIVIDUAL.includes(id)) {
                 const localPath = localStore.path
-                if (await cloudIsNewer(cloudPath, localPath)) {
+                if (await cloudIsNewer(localPath, modifiedDates[file.name])) {
                     await moveFileAsync(cloudPath, localPath)
                 }
                 return
@@ -269,16 +269,12 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
     // check any local instance not in cloud (exists only locally)
     const bibleNames = await readFolderAsync(biblesFolder)
     const localBibles = bibleNames.filter((key) => !cloudBibleNames.includes(key))
-    localBibles.forEach((fileName) => {
-        if (isDeleted("BIBLES", fileName)) {
+    await Promise.all(
+        localBibles.map(async (fileName) => {
             const localBiblePath = path.join(biblesFolder, fileName)
-            deleteFile(localBiblePath)
-            return
-        }
-
-        // new file
-        markAsCreated("BIBLES", fileName)
-    })
+            await createOrDeleteLocalFile("BIBLES", fileName, localBiblePath)
+        })
+    )
 
     if (readOnly) return finish()
 
@@ -328,17 +324,17 @@ export async function syncData(data: { id: SyncProviderId; churchId: string; tea
     function finish(success: boolean = true) {
         deleteFolder(EXTRACT_LOCATION)
         console.log("Sync completed!")
+        isNewDevice = false
         return { success, changedFiles }
     }
 }
 
-async function cloudIsNewer(cloudFilePath: any, localFilePath: any): Promise<boolean> {
-    const cloudStats = await getFileStatsAsync(cloudFilePath)
+async function cloudIsNewer(localFilePath: any, cloudDate: Date): Promise<boolean> {
     const localStats = await getFileStatsAsync(localFilePath)
-    if (!cloudStats) return false
+    if (!cloudDate) return false
     if (!localStats) return true
 
-    return cloudStats.mtime > localStats.mtime
+    return cloudDate.getTime() > localStats.mtime.getTime()
 }
 
 function getLocalOnlyKeys(cloudKeys: any, localKeys: any): string[] {
@@ -419,6 +415,29 @@ async function deleteUnusedZips(folderPath: string, excludeZip: string) {
     }
 }
 
+async function createOrDeleteLocalFile(id: ChangeId, fileName: string, localPath: string) {
+    // marked as deleted in general
+    if (isDeleted(id, fileName)) {
+        // marked as already deleted for this device
+        if (isDeletedLocally(id, fileName)) {
+            // already deleted locally
+            if (!(await doesPathExistAsync(localPath))) return
+
+            // revert if not deleted anymore
+            unmarkAsDeleted(id, fileName)
+            markAsCreated(id, fileName)
+            return
+        }
+
+        // delete local file
+        deleteFile(localPath)
+        return
+    }
+
+    // new file
+    markAsCreated(id, fileName)
+}
+
 // SYNC LOGIC
 // if not found locally, and marked as "deleted" in cloud: skip
 // if not found locally, and marked as "created" in cloud: download
@@ -426,9 +445,15 @@ async function deleteUnusedZips(folderPath: string, excludeZip: string) {
 // if found locally only, and not marked in cloud: mark as "created"
 // if found locally, but marked as "deleted" in cloud: delete locally
 // if found locally and in cloud: use newest version
+// if marked as deleted locally in cloud, but exists locally: unmark as deleted and mark as created
 
+type Changes = { version: string; devices: string[]; modified: { [key: string]: number }; deleted: { [key: string]: string[] }; created: { [key: string]: string[] } }
 const changes_name = "changes.json"
-let CHANGES: { devices: string[]; modified: { [key: string]: number }; deleted: { [key: string]: string[] }; created: { [key: string]: string[] } } = { devices: [], modified: {}, deleted: {}, created: {} }
+const version = "0.0.1"
+const DEFAULT_CHANGES: Changes = { version, devices: [], modified: {}, deleted: {}, created: {} }
+let CHANGES: Changes = clone(DEFAULT_CHANGES)
+let cloudChanges: Changes | null = null
+let isNewDevice = false
 
 // keep track of last changed time so we can know which devices to ignore eventually
 function getLatestChanges() {
@@ -465,12 +490,25 @@ function markAs(type: "deleted" | "created", instanceId: string) {
     }
 }
 
+function unmarkAsDeleted(storeId: ChangeId, key: string) {
+    if (!CHANGES.deleted) return
+
+    const instanceId = `${storeId}_${key}`
+    delete CHANGES.deleted[instanceId]
+}
+
 function isDeleted(storeId: ChangeId, key: string): boolean {
     const instanceId = `${storeId}_${key}`
     return !!CHANGES.deleted?.[instanceId]
 }
 
 function isCreated(storeId: ChangeId, key: string): boolean {
+    if (isNewDevice) return true
     const instanceId = `${storeId}_${key}`
     return !!CHANGES.created?.[instanceId]
+}
+
+function isDeletedLocally(storeId: ChangeId, key: string): boolean {
+    const instanceId = `${storeId}_${key}`
+    return !!cloudChanges?.deleted?.[instanceId]?.includes(deviceId)
 }
